+++ src/pykeen/evaluation/classification_evaluator.py
@@ -8,9 +8,8 @@ import numpy as np
 import torch
 
 from .evaluator import Evaluator, MetricResults
-from .rexmex_compat import classifier_annotator
 from ..constants import TARGET_TO_INDEX
-from ..metrics.utils import MetricAnnotation
+from ..metrics.classification import MetricAnnotation, classifier_annotator
 from ..typing import MappedTriples, Target
 
 __all__ = [
