+++ src/pykeen/metrics/classification.py
@@ -10,14 +10,37 @@ through :mod:`rexmex`.
 """
 
 import inspect
+from dataclasses import dataclass
+from typing import Callable, MutableMapping, Optional
 
+import numpy as np
 import rexmex.metrics.classification as rmc
 
-from ..metrics.utils import MetricAnnotator
+from .utils import ValueRange, construct_indicator
 
 __all__ = [
     "classifier_annotator",
 ]
+
+
+@dataclass
+class MetricAnnotation:
+    """Metadata about a classifier function."""
+
+    name: str
+    increasing: bool
+    value_range: ValueRange
+    description: str
+    link: str
+    func: Callable[[np.array, np.array], float]
+    binarize: bool
+
+    def score(self, y_true, y_score) -> float:
+        """Run the scoring function."""
+        if self.func is None:
+            raise ValueError
+        return self.func(y_true, construct_indicator(y_score=y_score, y_true=y_true) if self.binarize else y_score)
+
 
 #: Functions with the right signature in the :mod:`rexmex.metrics.classification` that are not themselves metrics
 EXCLUDE = {
@@ -39,6 +62,54 @@ DUPLICATE_CLASSIFIERS = {
     rmc.precision_score: rmc.positive_predictive_value,
     rmc.recall_score: rmc.true_positive_rate,
 }
+
+
+class MetricAnnotator:
+    """A class for annotating metric functions."""
+
+    metrics: MutableMapping[str, MetricAnnotation]
+
+    def __init__(self):
+        self.metrics = {}
+
+    def higher(self, func, **kwargs):
+        """Annotate a function where higher values are better."""
+        return self.add(func, increasing=True, **kwargs)
+
+    def lower(self, func, **kwargs):
+        """Annotate a function where lower values are better."""
+        return self.add(func, increasing=False, **kwargs)
+
+    def add(
+        self,
+        func,
+        *,
+        increasing: bool,
+        description: str,
+        link: str,
+        name: Optional[str] = None,
+        lower: Optional[float] = 0.0,
+        lower_inclusive: bool = True,
+        upper: Optional[float] = 1.0,
+        upper_inclusive: bool = True,
+        binarize: bool = False,
+    ):
+        """Annotate a function."""
+        self.metrics[func] = MetricAnnotation(
+            func=func,
+            binarize=binarize,
+            name=name or func.__name__.replace("_", " ").title(),
+            value_range=ValueRange(
+                lower=lower,
+                lower_inclusive=lower_inclusive,
+                upper=upper,
+                upper_inclusive=upper_inclusive,
+            ),
+            increasing=increasing,
+            description=description,
+            link=link,
+        )
+
 
 classifier_annotator = MetricAnnotator()
 classifier_annotator.higher(
