+++ src/pykeen/metrics/utils.py
@@ -3,7 +3,7 @@
 """Utilities for metrics."""
 
 from dataclasses import dataclass
-from typing import Callable, ClassVar, Collection, Iterable, MutableMapping, Optional
+from typing import ClassVar, Collection, Iterable, Optional
 
 import numpy as np
 from docdata import get_docdata
@@ -12,8 +12,6 @@ from ..utils import camel_to_snake
 
 __all__ = [
     "Metric",
-    "MetricAnnotation",
-    "MetricAnnotator",
     "construct_indicator",
     "ValueRange",
 ]
@@ -114,72 +112,6 @@ class Metric:
 
     def __str__(self) -> str:
         return f"{self.__class__.__name__}({', '.join(self._extra_repr())})"
-
-
-@dataclass
-class MetricAnnotation:
-    """Metadata about a classifier function."""
-
-    name: str
-    increasing: bool
-    value_range: ValueRange
-    description: str
-    link: str
-    func: Callable[[np.array, np.array], float]
-    binarize: bool
-
-    def score(self, y_true, y_score) -> float:
-        """Run the scoring function."""
-        if self.func is None:
-            raise ValueError
-        return self.func(y_true, construct_indicator(y_score=y_score, y_true=y_true) if self.binarize else y_score)
-
-
-class MetricAnnotator:
-    """A class for annotating metric functions."""
-
-    metrics: MutableMapping[str, MetricAnnotation]
-
-    def __init__(self):
-        self.metrics = {}
-
-    def higher(self, func, **kwargs):
-        """Annotate a function where higher values are better."""
-        return self.add(func, increasing=True, **kwargs)
-
-    def lower(self, func, **kwargs):
-        """Annotate a function where lower values are better."""
-        return self.add(func, increasing=False, **kwargs)
-
-    def add(
-        self,
-        func,
-        *,
-        increasing: bool,
-        description: str,
-        link: str,
-        name: Optional[str] = None,
-        lower: Optional[float] = 0.0,
-        lower_inclusive: bool = True,
-        upper: Optional[float] = 1.0,
-        upper_inclusive: bool = True,
-        binarize: bool = False,
-    ):
-        """Annotate a function."""
-        self.metrics[func] = MetricAnnotation(
-            func=func,
-            binarize=binarize,
-            name=name or func.__name__.replace("_", " ").title(),
-            value_range=ValueRange(
-                lower=lower,
-                lower_inclusive=lower_inclusive,
-                upper=upper,
-                upper_inclusive=upper_inclusive,
-            ),
-            increasing=increasing,
-            description=description,
-            link=link,
-        )
 
 
 def construct_indicator(*, y_score: np.ndarray, y_true: np.ndarray) -> np.ndarray:
